Clean up reducers into their own files
